Paused to rewrite the tours sub-project
Quite stable - more testing needed
Wiki reading list done. All callbacks handled (show on map, share, remove from list)
Wiki links directed to webview - image clicks inside webview are overridden and displayed in own image viewer
POI lists done

Removed directions handling and direct user to Google or other map applicaiton they have installed (might come back to own version at some point)

Todo:
Add image click listener to POI / wiki markers list. Maybe add 'more' button for extra options
Add map long click pop up menu

+++ app/src/main/java/me/carc/btown/Utils/MapZoom.java
@@ -27,7 +27,7 @@ public class MapZoom {
         double pixels = Math.sqrt((mapWidth * mapWidth) + (mapHeight * mapHeight));
         final double requiredMinimalGroundResolutionInMetersPerPixel = ((double) new GeoPoint(min.getLatitude(), min.getLongitude()).distanceToAsDouble(max)) / pixels;
         int zoom = calculateZoom(center.getLatitude(), requiredMinimalGroundResolutionInMetersPerPixel, tileProvider.getMaximumZoomLevel(), tileProvider.getMinimumZoomLevel());
-        controller.setZoom(zoom > 17 ? 17 : zoom-1);
+        controller.setZoom((double)(zoom > 17 ? 17 : zoom-1));
         controller.animateTo(center);
     }
